[feat] validate 클래스 추가
