+++ src/main/java/com/lof/lofserver/service/user/validate/UserValidate.java
@@ -1,0 +1,5 @@
+package com.lof.lofserver.service.user.validate;
+
+public interface UserValidate {
+    void validateNickname(String nickname);
+}
